requirements.txt
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.1
django-extensions==3.2.3
djangorestframework==3.14.0
duplicity==0.8.21
fasteners==0.14.1
fcm-django==2.0.1
firebase-admin==6.4.0
Flask==3.0.2
future==0.18.2
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
msgpack==1.0.8
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pycups==2.0.1
PyGetWindow==0.0.9
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.30
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.1
pywhatkit==5.4
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==3.0.1
wikipedia==1.4.0
xdg==5
xkit==0.0.0
zipp==1.0.0
